+++ modules/discord/example.config.json
@@ -1,0 +1,23 @@
+{
+    "token": "discord_bot_token",
+    "prefix": "++",
+    "tardyResponses": [
+		"Do better next time, okay?",
+		"I'm counting on you next time!",
+		"Oops. try again tomorrow!",
+		"No good. Perhaps tomorrow you will improve."
+    ],
+    "redemptionResponses": [
+		":)!!!!",
+		"I knew you could do it!",
+		":D!!!",
+		"<3<3<3<3"
+	],
+	"commands": [
+		"++help: List all commands \n",
+		"++master: Shows my author \n",
+		"++mrank: List the Morning scores\n",
+		"++timezone#: Set your difference in hours relative to EST (+/- hours)\n",
+		"++mytimezone: Show what I think your time is right now\n"
+	]
+}
