+++ modules/mysql/example.config.json
@@ -1,0 +1,6 @@
+{
+    "host" : "localhost",
+    "user" : "",
+    "password" : "",
+    "database" : "morning_log"
+}
