+++ modules/paperquotes/example.config.json
@@ -1,0 +1,4 @@
+{
+    "token": "paperquotes_api_token",
+    "start_date": "December 18, 2019"
+}
