+++ package.json
@@ -13,7 +13,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.21.0",
+    "axios": "^0.21.1",
     "cron": "^1.7.2",
     "discord.js": "^11.4.2",
     "express": "^4.17.1",
